requirements.txt
# WhisperX and its dependencies (versions from whisperx 3.7.4)
whisperx>=3.7.4

# Core dependencies (exact versions from whisperx pyproject.toml)
ctranslate2>=4.5.0
faster-whisper>=1.1.1
nltk>=3.9.1
numpy>=2.0.2,<2.1.0; python_version <'3.13'
numpy>=2.1.0,<2.3.0; python_version >='3.13'
onnxruntime>=1.19,<1.20.0; python_version <'3.10'
pandas>=2.2.3,<2.3.0
av<16.0.0
pyannote-audio>=3.3.2,<4.0.0
torch~=2.8.0
torchaudio~=2.8.0
transformers>=4.48.0
triton>=3.3.0; sys_platform == 'linux' and platform_machine == 'x86_64'

# Additional dependencies for the transcriber service
fastapi
uvicorn
pydantic
soundfile
librosa
python-multipart
